chore(si-pkg): break spec up into modules
