+++ lib/si-pkg/src/spec/variant.rs
@@ -1,0 +1,82 @@
+use derive_builder::Builder;
+use serde::{Deserialize, Serialize};
+use url::Url;
+
+use super::{LeafFunctionSpec, PropSpec, SpecError};
+
+#[derive(Builder, Clone, Debug, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+#[builder(build_fn(error = "SpecError"))]
+pub struct SchemaVariantSpec {
+    #[builder(setter(into))]
+    pub name: String,
+    #[builder(setter(into, strip_option), default)]
+    pub link: Option<Url>,
+    #[builder(setter(into, strip_option), default)]
+    pub color: Option<String>,
+
+    #[builder(private, default = "Self::default_domain()")]
+    pub domain: PropSpec,
+
+    #[builder(setter(each(name = "leaf_function"), into), default)]
+    pub leaf_functions: Vec<LeafFunctionSpec>,
+}
+
+impl SchemaVariantSpec {
+    pub fn builder() -> SchemaVariantSpecBuilder {
+        SchemaVariantSpecBuilder::default()
+    }
+}
+
+impl SchemaVariantSpecBuilder {
+    fn default_domain() -> PropSpec {
+        PropSpec::Object {
+            validations: None,
+            name: "domain".to_string(),
+            entries: vec![],
+        }
+    }
+
+    #[allow(unused_mut)]
+    pub fn try_link<V>(&mut self, value: V) -> Result<&mut Self, V::Error>
+    where
+        V: TryInto<Url>,
+    {
+        let converted: Url = value.try_into()?;
+        Ok(self.link(converted))
+    }
+
+    #[allow(unused_mut)]
+    pub fn prop(&mut self, item: impl Into<PropSpec>) -> &mut Self {
+        let converted: PropSpec = item.into();
+        match self.domain.get_or_insert_with(Self::default_domain) {
+            PropSpec::Object { entries, .. } => entries.push(converted),
+            invalid => unreachable!(
+                "domain prop is an object but was found to be: {:?}",
+                invalid
+            ),
+        };
+        self
+    }
+
+    #[allow(unused_mut)]
+    pub fn try_prop<I>(&mut self, item: I) -> Result<&mut Self, I::Error>
+    where
+        I: TryInto<PropSpec>,
+    {
+        let converted: PropSpec = item.try_into()?;
+        Ok(self.prop(converted))
+    }
+
+    #[allow(unused_mut)]
+    pub fn props(&mut self, value: Vec<PropSpec>) -> &mut Self {
+        match self.domain.get_or_insert_with(Self::default_domain) {
+            PropSpec::Object { entries, .. } => *entries = value,
+            invalid => unreachable!(
+                "domain prop is an object but was found to be: {:?}",
+                invalid
+            ),
+        };
+        self
+    }
+}
